+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 💥 Breaking changes
+- Moved `StrykerTasks` from `Candoumbe.Pipelines.Tools` to `Candoumbe.Pipelines.Tools.Stryker` namespace.
 
 ## [1.3.0] / 2025-11-12
 ### 🛠️ Fixes